+++ xrpl/core/binarycodec/types/xchain_attestation_batch.py
